:sparkles: payment presentation 구현
